requirements.txt
python-dotenv==1.0.1
tabulate==0.9.0
